# Conflicts:
#	MiAmusing/MiAmusing.jucer
#	MiAmusing/Source/AmusingScene.cpp
#	MiAmusing/Source/AmusingScene.h
#	MiAmusing/Source/AnimatedPolygon.h
#	MiAmusing/Source/AudioManager.cpp
#	MiAmusing/Source/CompletePolygon.cpp
#	MiAmusing/Source/MultiSceneCanvasManager.cpp
#	MiAmusing/Source/ReadingHead.cpp
#	MiAmusing/Source/editScene.cpp
#	MiAmusing/Source/editScene.h

+++ MiAmusing/Source/Presenter.cpp
@@ -16,7 +16,6 @@
 
 #include "JuceHeader.h"
 
-#include "AnimatedPolygon.h"
 #include "GraphicEvent.h"
 #include "MultiAreaEvent.h"
 #include "Cursors.h"
@@ -328,7 +327,7 @@ double Presenter::computeFrequency(std::shared_ptr<IEditableArea> area,double su
 	{
 		//return initFreq + (surface - areaToInitSurface[area]) * (minFreq - initFreq) / (maxSize - areaToInitSurface[area]);
 		//return std::exp(m*surface + p);
-		int idx = round((S - LUT::init) * LUT::size);
+		int idx = (int)round((S - LUT::init) * LUT::size);
 		return LUT::LUT[idx];//k + A * std::exp(m * S);
 	}
